# Saving disk space in conditioned GMFs calculations