--- a/openquake/calculators/event_based.py
+++ b/openquake/calculators/event_based.py
@@ -359,17 +359,17 @@ def starmap_from_rups(func, oq, full_lt, sitecol, dstore, save_tmp=None):
         G = len(cmaker.gsims)
         M = len(cmaker.imts)
         N = len(computer.sitecol)
-        size = 3 * G * M * N * N * 8  # sig, tau, phi
+        size = 2 * G * M * N * N * 8  # tau, phi
         logging.info('Building %s of mean_covs', humansize(size))
         if size > float(config.memory.conditioned_gmf_gb) * 1024**3:
             raise ValueError(
                 f'The calculation is too large: {G=}, {M=}, {N=}. '
                 'You must reduce the number of sites i.e. enlarge '
                 'region_grid_spacing)')
-        mean_covs = computer.get_mean_covs()
+        mea, _, tau, phi = computer.get_mean_covs()
         del proxy.geom  # to reduce data transfer
     else:
-        mean_covs = ()
+        mea, tau, phi = None, None, None
     dstore.swmr_on()
     smap = parallel.Starmap(func, h5=dstore.hdf5)
     if save_tmp:
@@ -380,7 +380,7 @@ def starmap_from_rups(func, oq, full_lt, sitecol, dstore, save_tmp=None):
         rlzs_by_gsim = full_lt.get_rlzs_by_gsim(trt_smr)
         cmaker = ContextMaker(trt, rlzs_by_gsim, oq, extraparams=extra)
         cmaker.min_mag = getdefault(oq.minimum_magnitude, trt)
-        cmaker.mean_covs = mean_covs
+        cmaker.mea_tau_phi = mea, tau, phi
         for block in block_splitter(proxies, totw, rup_weight):
             args = block, cmaker, (station_data, station_sites), dstore
             smap.submit(args)